+++ packages/storefront-client/vite.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from "vite";
 import dts from "vite-plugin-dts";
-// import pkg from "./package.json" assert { type: "json" };
+import pkg from "./package.json" assert { type: "json" };
 
 export default defineConfig({
   server: {
@@ -9,10 +9,29 @@ export default defineConfig({
   build: {
     lib: {
       name: "Shopify.StorefrontClient",
-      entry: "./src/index.ts",
-      formats: ["es", "umd", "iife"],
+      entry: {
+        index: "./src/index.ts",
+      },
+      formats: ["es", "cjs"],
     },
-    rollupOptions: {},
+    rollupOptions: {
+      external: [
+        "solid-js",
+        "@tanstack/solid-query",
+        "@tanstack/query-sync-storage-persister",
+        "@tanstack/solid-query-persist-client",
+        "lz-string",
+      ],
+      output: {
+        name: "Shopify.StorefrontClient",
+        // paths: (id) => {
+        //   const version = pkg.dependencies[id]?.replace("^", "@");
+        //   if (!version) return id;
+        //   const url = `https://cdn.jsdelivr.net/npm/${id}${version}/+esm`;
+        //   return url;
+        // },
+      },
+    },
     target: "esnext", // transpile as little as possible
   },
   plugins: [dts()], // emit TS declaration files