+++ packages/web-components/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @bryt-designs/web-components
 
+## 0.4.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @bryt-designs/storefront-client@0.4.1
+  - @bryt-designs/predictive-search@0.3.1
+
 ## 0.4.3
 
 ### Patch Changes